Fix issues with merging
